+++ src/options/args.rs
@@ -5,41 +5,28 @@
 
 // TODO: New sections are misaligned! See if we can get that fixed.
 
-use std::cmp::Ordering;
+use std::path::PathBuf;
 
-use clap::*;
+use clap::{builder::PossibleValue, *};
 use indoc::indoc;
 
-pub fn get_matches() -> ArgMatches {
-    build_app().get_matches()
-}
+const TEMPLATE: &str = indoc! {
+    "{name} {version}
+    {author}
 
-/// Returns an [`Ordering`] for two [`Arg`] values.
-///
-/// Note this assumes that they both have a _long_ name, and will
-/// panic if either are missing!
-fn sort_args(a: &Arg, b: &Arg) -> Ordering {
-    let a = a.get_long().unwrap();
-    let b = b.get_long().unwrap();
+    {about}
 
-    a.cmp(b)
-}
+    {usage-heading} {usage}
 
-/// Create an array of [`Arg`] values. If there is more than one value, then
-/// they will be sorted by their long name. Note this sort will panic if
-/// any [`Arg`] does not have a long name!
-macro_rules! args {
-    ( $arg:expr $(,)?) => {
-        [$arg]
-    };
-    ( $( $arg:expr ),+ $(,)? ) => {
-        {
-            let mut args = [ $( $arg, )* ];
-            args.sort_unstable_by(sort_args);
-            args
-        }
-    };
-}
+    {all-args}"
+};
+
+const USAGE: &str = "btm [OPTIONS]";
+
+const VERSION: &str = match option_env!("NIGHTLY_VERSION") {
+    Some(nightly_version) => nightly_version,
+    None => crate_version!(),
+};
 
 const CHART_WIDGET_POSITIONS: [&str; 9] = [
     "none",
@@ -53,57 +40,104 @@ const CHART_WIDGET_POSITIONS: [&str; 9] = [
     "bottom-right",
 ];
 
-fn general_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("General Options");
+/// Represents the arguments that can be passed in to bottom.
+#[derive(Parser, Debug)]
+#[command(
+    name = crate_name!(),
+    version = VERSION,
+    author = crate_authors!(),
+    about = crate_description!(),
+    disable_help_flag = true,
+    disable_version_flag = true,
+    color = ColorChoice::Auto,
+    help_template = TEMPLATE,
+    override_usage = USAGE,
+)]
+pub struct BottomArgs {
+    #[command(flatten)]
+    pub general: GeneralArgs,
 
-    let autohide_time = Arg::new("autohide_time")
-        .long("autohide_time")
-        .action(ArgAction::SetTrue)
-        .help("Temporarily shows the time scale in graphs.")
-        .long_help(
-            "Automatically hides the time scale in graphs after being shown for a brief moment when zoomed \
-            in/out. If time is disabled via --hide_time then this will have no effect."
-        );
+    #[command(flatten)]
+    pub process: ProcessArgs,
 
-    let basic = Arg::new("basic")
-        .short('b')
-        .long("basic")
-        .action(ArgAction::SetTrue)
-        .help("Hides graphs and uses a more basic look.")
-        .long_help("Hides graphs and uses a more basic look, largely inspired by htop's design.");
+    #[command(flatten)]
+    pub temperature: TemperatureArgs,
 
-    let config_location = Arg::new("config_location")
-        .short('C')
-        .long("config")
-        .action(ArgAction::Set)
-        .value_name("CONFIG PATH")
-        .help("Sets the location of the config file.")
-        .long_help(
-            "Sets the location of the config file. Expects a config file in the TOML format. \
-            If it doesn't exist, a default config file is created at the path. If no path is provided, \
-            the default config location will be used."
-        )
-        .value_hint(ValueHint::AnyPath);
+    #[command(flatten)]
+    pub cpu: CpuArgs,
 
-    let default_time_value = Arg::new("default_time_value")
-        .short('t')
-        .long("default_time_value")
-        .action(ArgAction::Set)
-        .value_name("TIME")
-        .help("Default time value for graphs.")
-        .long_help(
-            "Default time value for graphs. Either a number in milliseconds or a 'human duration' \
-            (e.g. 60s, 10m). Defaults to 60s, must be at least 30s.",
-        );
+    #[command(flatten)]
+    pub memory: MemoryArgs,
+
+    #[command(flatten)]
+    pub network: NetworkArgs,
+
+    #[cfg(feature = "battery")]
+    #[command(flatten)]
+    pub battery: BatteryArgs,
+
+    #[cfg(feature = "gpu")]
+    #[command(flatten)]
+    pub gpu: GpuArgs,
+
+    #[command(flatten)]
+    pub style: StyleArgs,
+
+    #[command(flatten)]
+    pub other: OtherArgs,
+}
+
+/// General arguments/config options.
+#[derive(Args, Clone, Debug)]
+#[command(next_help_heading = "General Options", rename_all = "snake_case")]
+pub struct GeneralArgs {
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Temporarily shows the time scale in graphs.",
+        long = "Automatically hides the time scale in graphs after being shown for a brief moment when zoomed \
+                in/out. If time is disabled using --hide_time then this will have no effect."
+    )]
+    pub autohide_time: bool,
+
+    #[arg(
+        short = 'b',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Hides graphs and uses a more basic look.",
+        long_help = "Hides graphs and uses a more basic look, largely inspired by htop's design."
+    )]
+    pub basic: bool,
+
+    #[arg(
+        short = 'C',
+        long,
+        value_name = "PATH",
+        value_hint = ValueHint::AnyPath,
+        help = "Sets the location of the config file.",
+        long_help = "Sets the location of the config file. Expects a config file in the TOML format. \
+                    If it doesn't exist, a default config file is created at the path. If no path is provided, \
+                    the default config location will be used."
+    )]
+    pub config_location: Option<PathBuf>,
+
+    #[arg(
+        short = 't',
+        long,
+        value_name = "TIME",
+        help = "Default time value for graphs.",
+        long_help = "Default time value for graphs. Either a number in milliseconds or a 'human duration' \
+                    (e.g. 60s, 10m). Defaults to 60s, must be at least 30s."
+    )]
+    pub default_time_value: Option<String>,
 
     // TODO: Charts are broken in the manpage
-    let default_widget_count = Arg::new("default_widget_count")
-        .long("default_widget_count")
-        .action(ArgAction::Set)
-        .requires_all(["default_widget_type"])
-        .value_name("N")
-        .help("Sets the N'th selected widget type as the default.")
-        .long_help(indoc! {
+    #[arg(
+        long,
+        requires_all = ["default_widget_type"],
+        value_name = "N",
+        help = "Sets the N'th selected widget type as the default.",
+        long_help = indoc! {
             "Sets the N'th selected widget type to use as the default widget. Requires 'default_widget_type' to also be \
             set, and defaults to 1.
 
@@ -117,14 +151,15 @@ fn general_args(cmd: Command) -> Command {
             And we set our default widget type to 'CPU'. If we set '--default_widget_count 1', then it would use the \
             CPU (1) as the default widget. If we set '--default_widget_count 3', it would use CPU (3) as the default \
             instead."
-        });
+        }
+    )]
+    pub default_widget_count: Option<u64>,
 
-    let default_widget_type = Arg::new("default_widget_type")
-        .long("default_widget_type")
-        .action(ArgAction::Set)
-        .value_name("WIDGET")
-        .help("Sets the default widget type, use `--help` for info.")
-        .long_help(indoc!{
+    #[arg(
+        long,
+        value_name = "WIDGET",
+        help = "Sets the default widget type. Use --help for more info.",
+        long_help = indoc!{
             "Sets which widget type to use as the default widget. For the default \
             layout, this defaults to the 'process' widget. For a custom layout, it defaults \
             to the first widget it sees.
@@ -136,9 +171,9 @@ fn general_args(cmd: Command) -> Command {
             | Process | CPU (3) | Temperature | CPU (4) |
             +---------+---------+-------------+---------+
 
-            Setting '--default_widget_type temperature' will make the temperature widget selected by default."
-        })
-        .value_parser([
+            Then, setting '--default_widget_type temperature' will make the temperature widget selected by default."
+        },
+        value_parser = [
             "cpu",
             "mem",
             "net",
@@ -153,359 +188,363 @@ fn general_args(cmd: Command) -> Command {
             "batt",
             #[cfg(feature = "battery")]
             "battery",
-        ]);
+        ],
+    )]
+    pub default_widget_type: Option<String>,
 
-    let disable_click = Arg::new("disable_click")
-        .long("disable_click")
-        .action(ArgAction::SetTrue)
-        .help("Disables mouse clicks.")
-        .long_help("Disables mouse clicks from interacting with bottom.");
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Disables mouse clicks.",
+        long_help = "Disables mouse clicks from interacting with bottom."
+    )]
+    pub disable_click: bool,
 
     // TODO: Change this to accept a string with the type of marker.
-    let dot_marker = Arg::new("dot_marker")
-        .short('m')
-        .long("dot_marker")
-        .action(ArgAction::SetTrue)
-        .help("Uses a dot marker for graphs.")
-        .long_help("Uses a dot marker for graphs as opposed to the default braille marker.");
+    #[arg(
+        short = 'm',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Uses a dot marker for graphs.",
+        long_help = "Uses a dot marker for graphs as opposed to the default braille marker."
+    )]
+    pub dot_marker: bool,
 
-    let expanded = Arg::new("expanded")
-        .short('e')
-        .long("expanded")
-        .action(ArgAction::SetTrue)
-        .help("Expand the default widget upon starting the app.")
-        .long_help("Expand the default widget upon starting the app. This flag has no effect in basic mode (--basic).");
+    #[arg(
+        short = 'e',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Expand the default widget upon starting the app.",
+        long_help = "Expand the default widget upon starting the app. This flag has no effect in basic mode (--basic)."
+    )]
+    pub expanded: bool,
 
-    let hide_table_gap = Arg::new("hide_table_gap")
-        .long("hide_table_gap")
-        .action(ArgAction::SetTrue)
-        .help("Hides spacing between table headers and entries.");
+    #[arg(long, action = ArgAction::SetTrue, help = "Hides spacing between table headers and entries.")]
+    pub hide_table_gap: bool,
 
-    let hide_time = Arg::new("hide_time")
-        .long("hide_time")
-        .action(ArgAction::SetTrue)
-        .help("Hides the time scale from being shown.");
+    #[arg(long, action = ArgAction::SetTrue, help = "Hides the time scale from being shown.")]
+    pub hide_time: bool,
 
-    let rate = Arg::new("rate")
-        .short('r')
-        .long("rate")
-        .action(ArgAction::Set)
-        .value_name("TIME")
-        .help("Sets how often data is refreshed.")
-        .long_help(
-            "Sets how often data is refreshed. Either a number in milliseconds or a 'human duration' \
-            (e.g. 1s, 1m). Defaults to 1s, must be at least 250ms. Smaller values may result in \
-            higher system resource usage."
-        );
+    #[arg(
+        short = 'r',
+        long,
+        value_name = "TIME",
+        help = "Sets how often data is refreshed.",
+        long_help = "Sets how often data is refreshed. Either a number in milliseconds or a 'human duration' \
+                    (e.g. 1s, 1m). Defaults to 1s, must be at least 250ms. Smaller values may result in \
+                    higher system resource usage."
+    )]
+    pub rate: Option<String>,
 
-    // TODO: Unify how we do defaults.
-    let retention = Arg::new("retention")
-        .long("retention")
-        .action(ArgAction::Set)
-        .value_name("TIME")
-        .help("How far back data will be stored up to.")
-        .long_help(
-            "How far back data will be stored up to. Either a number in milliseconds or a 'human duration' \
-            (e.g. 10m, 1h). Defaults to 10 minutes, and must be at least  1 minute. Larger values \
-            may result in higher memory usage."
-        );
+    #[arg(
+        long,
+        value_name = "TIME",
+        help = "How far back data will be stored up to.",
+        long_help = "How far back data will be stored up to. Either a number in milliseconds or a 'human duration' \
+                    (e.g. 10m, 1h). Defaults to 10 minutes, and must be at least  1 minute. Larger values \
+                    may result in higher memory usage."
+    )]
+    pub retention: Option<String>,
 
-    let show_table_scroll_position = Arg::new("show_table_scroll_position")
-        .long("show_table_scroll_position")
-        .action(ArgAction::SetTrue)
-        .help("Shows the scroll position tracker in table widgets.")
-        .long_help("Shows the list scroll position tracker in the widget title for table widgets.");
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Shows the list scroll position tracker in the widget title for table widgets."
+    )]
+    pub show_table_scroll_position: bool,
 
-    let time_delta = Arg::new("time_delta")
-        .short('d')
-        .long("time_delta")
-        .action(ArgAction::Set)
-        .value_name("TIME")
-        .help("The amount of time changed upon zooming.")
-        .long_help(
-            "The amount of time changed when zooming in/out. Takes a number in \
-            milliseconds or a human duration (e.g. 30s). The minimum is 1s, and \
-            defaults to 15s.",
-        );
-
-    cmd.args(args![
-        autohide_time,
-        basic,
-        config_location,
-        default_widget_count,
-        default_time_value,
-        default_widget_type,
-        disable_click,
-        dot_marker,
-        expanded,
-        hide_table_gap,
-        hide_time,
-        rate,
-        retention,
-        show_table_scroll_position,
-        time_delta,
-    ])
+    #[arg(
+        short = 'd',
+        long,
+        value_name = "TIME",
+        help = "The amount of time changed upon zooming.",
+        long_help = "The amount of time changed when zooming in/out. Takes a number in \
+                    milliseconds or a human duration (e.g. 30s). The minimum is 1s, and \
+                    defaults to 15s."
+    )]
+    pub time_delta: Option<String>,
 }
 
-fn process_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("Process Options");
-
-    let case_sensitive = Arg::new("case_sensitive")
-        .short('S')
-        .long("case_sensitive")
-        .action(ArgAction::SetTrue)
-        .help("Enables case sensitivity by default.")
-        .long_help("Enables case sensitivity by default when searching for a process.");
+/// Process arguments/config options.
+#[derive(Args, Clone, Debug, Default)]
+#[command(next_help_heading = "Process Options", rename_all = "snake_case")]
+pub struct ProcessArgs {
+    #[arg(
+        short = 'S',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Enables case sensitivity by default.",
+        long_help = "Enables case sensitivity by default when searching for a process."
+    )]
+    pub case_sensitive: bool,
 
     // TODO: Rename this.
-    let current_usage = Arg::new("current_usage")
-        .short('u')
-        .long("current_usage")
-        .action(ArgAction::SetTrue)
-        .help("Calculates process CPU usage as a percentage of current usage rather than total usage.");
+    #[arg(
+        short = 'u',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Calculates process CPU usage as a percentage of current usage rather than total usage."
+    )]
+    pub current_usage: bool,
 
     // TODO: Disable this on Windows?
-    let disable_advanced_kill = Arg::new("disable_advanced_kill")
-        .long("disable_advanced_kill")
-        .action(ArgAction::SetTrue)
-        .help("Hides additional stopping options Unix-like systems.")
-        .long_help(
-            "Hides additional stopping options Unix-like systems. Signal 15 (TERM) will be sent when \
-            stopping a process.",
-        );
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Hides additional stopping options Unix-like systems.",
+        long_help = "Hides additional stopping options Unix-like systems. Signal 15 (TERM) will be sent when \
+                    stopping a process."
+    )]
+    pub disable_advanced_kill: bool,
 
-    let group_processes = Arg::new("group_processes")
-        .short('g')
-        .long("group_processes")
-        .action(ArgAction::SetTrue)
-        .help("Groups processes with the same name by default.");
+    #[arg(
+        short = 'g',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Groups processes with the same name by default."
+    )]
+    pub group_processes: bool,
 
-    let mem_as_value = Arg::new("mem_as_value")
-        .long("mem_as_value")
-        .action(ArgAction::SetTrue)
-        .help("Defaults to showing process memory usage by value.")
-        .long_help("Defaults to showing process memory usage by value. Otherwise, it defaults to showing it by percentage.");
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Defaults to showing process memory usage by value.",
+        long_help = "Defaults to showing process memory usage by value. Otherwise, it defaults to showing it by percentage."
+    )]
+    pub process_memory_as_value: bool,
 
-    let process_command = Arg::new("process_command")
-        .long("process_command")
-        .action(ArgAction::SetTrue)
-        .help("Shows the full command name instead of the process name by default.");
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Shows the full command name instead of the process name by default."
+    )]
+    pub process_command: bool,
 
-    let regex = Arg::new("regex")
-        .short('R')
-        .long("regex")
-        .action(ArgAction::SetTrue)
-        .help("Enables regex by default while searching.");
+    #[arg(short = 'R', long, action = ArgAction::SetTrue, help = "Enables regex by default while searching.")]
+    pub regex: bool,
 
-    let tree = Arg::new("tree")
-        .short('T')
-        .long("tree")
-        .action(ArgAction::SetTrue)
-        .help("Makes the process widget use tree mode by default.");
+    #[arg(
+        short = 'T',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Makes the process widget use tree mode by default."
+    )]
+    pub tree: bool,
 
-    let unnormalized_cpu = Arg::new("unnormalized_cpu")
-        .short('n')
-        .long("unnormalized_cpu")
-        .action(ArgAction::SetTrue)
-        .help("Show process CPU% usage without averaging over the number of CPU cores.");
+    #[arg(
+        short = 'n',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Show process CPU% usage without averaging over the number of CPU cores."
+    )]
+    pub unnormalized_cpu: bool,
 
-    let whole_word = Arg::new("whole_word")
-        .short('W')
-        .long("whole_word")
-        .action(ArgAction::SetTrue)
-        .help("Enables whole-word matching by default while searching.");
-
-    let args = args![
-        case_sensitive,
-        current_usage,
-        disable_advanced_kill,
-        group_processes,
-        mem_as_value,
-        process_command,
-        regex,
-        tree,
-        unnormalized_cpu,
-        whole_word,
-    ];
-
-    cmd.args(args)
+    #[arg(
+        short = 'W',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Enables whole-word matching by default while searching."
+    )]
+    pub whole_word: bool,
 }
 
-fn temperature_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("Temperature Options");
+/// Temperature arguments/config options.
+#[derive(Args, Clone, Debug, Default)]
+#[command(next_help_heading = "Temperature Options", rename_all = "snake_case")]
+#[group(id = "temperature_unit", multiple = false)]
+pub struct TemperatureArgs {
+    #[arg(
+        short = 'c',
+        long,
+        action = ArgAction::SetTrue,
+        group = "temperature_unit",
+        help = "Use Celsius as the temperature unit. Default.",
+        long_help = "Use Celsius as the temperature unit. This is the default option."
+    )]
+    pub celsius: bool,
 
-    let celsius = Arg::new("celsius")
-        .short('c')
-        .long("celsius")
-        .action(ArgAction::SetTrue)
-        .help("Use Celsius as the temperature unit. Default.")
-        .long_help("Use Celsius as the temperature unit. This is the default option.");
+    #[arg(
+        short = 'f',
+        long,
+        action = ArgAction::SetTrue,
+        group = "temperature_unit",
+        help = "Use Fahrenheit as the temperature unit."
+    )]
+    pub fahrenheit: bool,
 
-    let fahrenheit = Arg::new("fahrenheit")
-        .short('f')
-        .long("fahrenheit")
-        .action(ArgAction::SetTrue)
-        .help("Use Fahrenheit as the temperature unit.");
-
-    let kelvin = Arg::new("kelvin")
-        .short('k')
-        .long("kelvin")
-        .action(ArgAction::SetTrue)
-        .help("Use Kelvin as the temperature unit.");
-
-    let temperature_group = ArgGroup::new("TEMPERATURE_TYPE").args([
-        celsius.get_id(),
-        fahrenheit.get_id(),
-        kelvin.get_id(),
-    ]);
-
-    cmd.args(args![celsius, fahrenheit, kelvin])
-        .group(temperature_group)
+    #[arg(
+        short = 'k',
+        long,
+        action = ArgAction::SetTrue,
+        group = "temperature_unit",
+        help = "Use Kelvin as the temperature unit."
+    )]
+    pub kelvin: bool,
 }
 
-fn cpu_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("CPU Options");
-
-    // let default_cpu_entry = Arg::new("");
-
-    let hide_avg_cpu = Arg::new("hide_avg_cpu")
-        .short('a')
-        .long("hide_avg_cpu")
-        .action(ArgAction::SetTrue)
-        .help("Hides the average CPU usage entry.");
-
-    let cpu_left_legend = Arg::new("cpu_left_legend")
-        .long("cpu_left_legend")
-        .action(ArgAction::SetTrue)
-        .help("Puts the CPU chart legend on the left side.");
-
-    cmd.args(args![hide_avg_cpu, cpu_left_legend])
+/// The default selection of the CPU widget. If the given selection is invalid,
+/// we will fall back to all.
+#[derive(Clone, Copy, Debug, Default)]
+pub enum CpuDefault {
+    #[default]
+    All,
+    Average,
 }
 
-fn mem_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("Memory Options");
+impl ValueEnum for CpuDefault {
+    fn value_variants<'a>() -> &'a [Self] {
+        &[CpuDefault::All, CpuDefault::Average]
+    }
 
-    let memory_legend = Arg::new("memory_legend")
-        .long("memory_legend")
-        .action(ArgAction::Set)
-        .value_name("POSITION")
-        .ignore_case(true)
-        .help("Where to place the legend for the memory chart widget.")
-        .value_parser(CHART_WIDGET_POSITIONS);
+    fn to_possible_value(&self) -> Option<PossibleValue> {
+        match self {
+            CpuDefault::All => Some(PossibleValue::new("all")),
+            CpuDefault::Average => Some(PossibleValue::new("avg").alias("average")),
+        }
+    }
+}
+
+/// CPU arguments/config options.
+#[derive(Args, Clone, Debug, Default)]
+#[command(next_help_heading = "CPU Options", rename_all = "snake_case")]
+pub struct CpuArgs {
+    #[arg(
+        long,
+        help = "Sets which CPU entry type is selected by default.",
+        value_name = "ENTRY",
+        value_parser = value_parser!(CpuDefault),
+        default_value = "all"
+    )]
+    pub default_cpu_entry: CpuDefault,
+
+    #[arg(short = 'a', long, action = ArgAction::SetTrue, help = "Hides the average CPU usage entry.")]
+    pub hide_avg_cpu: bool,
+
+    // TODO: Maybe rename this or fix this? Should this apply to all "left legends"?
+    #[arg(
+        short = 'l',
+        long,
+        action = ArgAction::SetTrue,
+        help = "Puts the CPU chart legend on the left side."
+    )]
+    pub cpu_left_legend: bool,
+}
+
+/// Memory argument/config options.
+#[derive(Args, Clone, Debug, Default)]
+#[command(next_help_heading = "Memory Options", rename_all = "snake_case")]
+pub struct MemoryArgs {
+    #[arg(
+        long,
+        value_parser = CHART_WIDGET_POSITIONS,
+        value_name = "POSITION",
+        ignore_case = true,
+        help = "Where to place the legend for the memory chart widget.",
+    )]
+    pub memory_legend: Option<String>,
 
     #[cfg(not(target_os = "windows"))]
-    {
-        let enable_cache_memory = Arg::new("enable_cache_memory")
-            .long("enable_cache_memory")
-            .action(ArgAction::SetTrue)
-            .help("Enable collecting and displaying cache and buffer memory.");
-
-        cmd.args(args![enable_cache_memory, memory_legend])
-    }
-    #[cfg(target_os = "windows")]
-    {
-        cmd.arg(memory_legend)
-    }
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Enables collecting and displaying cache and buffer memory."
+    )]
+    pub enable_cache_memory: bool,
 }
 
-fn network_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("Network Options");
+/// Network arguments/config options.
+#[derive(Args, Clone, Debug, Default)]
+#[command(next_help_heading = "Network Options", rename_all = "snake_case")]
+pub struct NetworkArgs {
+    #[arg(
+        long,
+        value_parser = CHART_WIDGET_POSITIONS,
+        value_name = "POSITION",
+        ignore_case = true,
+        help = "Where to place the legend for the network chart widget.",
+    )]
+    pub network_legend: Option<String>,
 
-    let network_legend = Arg::new("network_legend")
-        .long("network_legend")
-        .action(ArgAction::Set)
-        .value_name("POSITION")
-        .ignore_case(true)
-        .help("Where to place the legend for the network chart widget.")
-        .value_parser(CHART_WIDGET_POSITIONS);
+    // TODO: Rename some of these to remove the network prefix for serde.
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Displays the network widget using bytes.",
+        long_help = "Displays the network widget using bytes. Defaults to bits."
+    )]
+    pub network_use_bytes: bool,
 
-    let network_use_bytes = Arg::new("network_use_bytes")
-        .long("network_use_bytes")
-        .action(ArgAction::SetTrue)
-        .help("Displays the network widget using bytes.")
-        .long_help("Displays the network widget using bytes. Defaults to bits.");
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Displays the network widget with binary prefixes.",
+        long_help = "Displays the network widget with binary prefixes (e.g. kibibits, mebibits) rather than a decimal \
+                    prefixes (e.g. kilobits, megabits). Defaults to decimal prefixes."
+    )]
+    pub network_use_binary_prefix: bool,
 
-    let network_use_binary_prefix = Arg::new("network_use_binary_prefix")
-        .long("network_use_binary_prefix")
-        .action(ArgAction::SetTrue)
-        .help("Displays the network widget with binary prefixes.")
-        .long_help(
-            "Displays the network widget with binary prefixes (e.g. kibibits, mebibits) rather than a decimal \
-            prefixes (e.g. kilobits, megabits). Defaults to decimal prefixes."
-        );
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Displays the network widget with a log scale.",
+        long_help = "Displays the network widget with a log scale. Defaults to a non-log scale."
+    )]
+    pub network_use_log: bool,
 
-    let network_use_log = Arg::new("network_use_log")
-        .long("network_use_log")
-        .action(ArgAction::SetTrue)
-        .help("Displays the network widget with a log scale.")
-        .long_help("Displays the network widget with a log scale. Defaults to a non-log scale.");
-
-    // TODO: Change this to be configured as network graph type?
-    let use_old_network_legend = Arg::new("use_old_network_legend")
-        .long("use_old_network_legend")
-        .action(ArgAction::SetTrue)
-        .help("(DEPRECATED) Uses a separated network legend.")
-        .long_help("(DEPRECATED) Uses separated network widget legend. This display is not tested and may be broken.");
-
-    cmd.args(args![
-        network_legend,
-        network_use_bytes,
-        network_use_log,
-        network_use_binary_prefix,
-        use_old_network_legend,
-    ])
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "(DEPRECATED) Uses a separate network legend.",
+        long_help = "(DEPRECATED) Uses separate network widget legend. This display is not tested and may be broken."
+    )]
+    pub use_old_network_legend: bool,
 }
 
+/// Battery arguments/config options.
 #[cfg(feature = "battery")]
-fn battery_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("Battery Options");
-
-    let battery = Arg::new("battery")
-        .long("battery")
-        .action(ArgAction::SetTrue)
-        .help("Shows the battery widget in non-custom layouts.")
-        .long_help(
-            "Shows the battery widget in default or basic mode, if there is as battery available. This \
-            has no effect on custom layouts; if the battery widget is desired for a custom layout, explicitly \
-            specify it."
-        );
-
-    cmd.arg(battery)
+#[derive(Args, Clone, Debug, Default)]
+#[command(next_help_heading = "Battery Options", rename_all = "snake_case")]
+pub struct BatteryArgs {
+    #[arg(
+        long,
+        action = ArgAction::SetTrue,
+        help = "Shows the battery widget in non-custom layouts.",
+        long_help = "Shows the battery widget in default or basic mode, if there is as battery available. This \
+                    has no effect on custom layouts; if the battery widget is desired for a custom layout, explicitly \
+                    specify it."
+    )]
+    pub battery: bool,
 }
 
+/// GPU arguments/config options.
 #[cfg(feature = "gpu")]
-fn gpu_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("GPU Options");
-
-    let enable_gpu = Arg::new("enable_gpu")
-        .long("enable_gpu")
-        .action(ArgAction::SetTrue)
-        .help("Enable collecting and displaying GPU usage.");
-
-    cmd.arg(enable_gpu)
+#[derive(Args, Clone, Debug, Default)]
+#[command(next_help_heading = "GPU Options", rename_all = "snake_case")]
+pub struct GpuArgs {
+    #[arg(long, action = ArgAction::SetTrue, help = "Enable collecting and displaying GPU usage.")]
+    pub enable_gpu: bool,
 }
 
-fn style_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("Style Options");
-
-    // TODO: File an issue with manpage, it cannot render charts correctly.
-    let color = Arg::new("color")
-        .long("color")
-        .action(ArgAction::Set)
-        .value_name("SCHEME")
-        .value_parser([
+/// Style arguments/config options.
+#[derive(Args, Clone, Debug, Default)]
+#[command(next_help_heading = "Style Options", rename_all = "snake_case")]
+pub struct StyleArgs {
+    #[arg(
+        long,
+        value_name = "SCHEME",
+        value_parser = [
             "default",
             "default-light",
             "gruvbox",
             "gruvbox-light",
             "nord",
             "nord-light",
-        ])
-        .hide_possible_values(true)
-        .help(indoc! {
+        ],
+        hide_possible_values = true,
+        help = indoc! {
             "Use a color scheme, use `--help` for info on the colors. [possible values: default, default-light, gruvbox, gruvbox-light, nord, nord-light]",
-        })
-        .long_help(indoc! {
+        },
+        long_help = indoc! {
             "Use a pre-defined color scheme. Currently supported values are:
             - default
             - default-light (default but adjusted for lighter backgrounds)
@@ -513,72 +552,31 @@ fn style_args(cmd: Command) -> Command {
             - gruvbox-light (gruvbox but adjusted for lighter backgrounds)
             - nord          (an arctic, north-bluish color palette)
             - nord-light    (nord but adjusted for lighter backgrounds)"
-        });
-
-    cmd.arg(color)
+        }
+    )]
+    pub color: Option<String>,
 }
 
-fn other_args(cmd: Command) -> Command {
-    let cmd = cmd.next_help_heading("Other Options");
+/// Other arguments. This just handle options that are for help/version displaying.
+#[derive(Args, Clone, Debug)]
+#[command(next_help_heading = "Other Options", rename_all = "snake_case")]
+pub struct OtherArgs {
+    #[arg(short = 'h', long, action = ArgAction::Help, help = "Prints help info (for more details use `--help`.")]
+    help: (),
 
-    let help = Arg::new("help")
-        .short('h')
-        .long("help")
-        .action(ArgAction::Help)
-        .help("Prints help info (for more details use `--help`.");
-
-    let version = Arg::new("version")
-        .short('V')
-        .long("version")
-        .action(ArgAction::Version)
-        .help("Prints version information.");
-
-    cmd.args([help, version])
+    #[arg(short = 'v', long, action = ArgAction::Version, help = "Prints version information.")]
+    version: (),
 }
 
-pub fn build_app() -> Command {
-    const TEMPLATE: &str = indoc! {
-        "{name} {version}
-        {author}
-    
-        {about}
-    
-        {usage-heading} {usage}
-    
-        {all-args}"
-    };
-    const USAGE: &str = "btm [OPTIONS]";
-    const VERSION: &str = match option_env!("NIGHTLY_VERSION") {
-        Some(nightly_version) => nightly_version,
-        None => crate_version!(),
-    };
+/// Returns a [`BottomArgs`].
+pub fn get_args() -> BottomArgs {
+    BottomArgs::parse()
+}
 
-    let cmd = Command::new(crate_name!())
-        .author(crate_authors!())
-        .about(crate_description!())
-        .disable_help_flag(true)
-        .disable_version_flag(true)
-        .color(ColorChoice::Auto)
-        .help_template(TEMPLATE)
-        .override_usage(USAGE)
-        .version(VERSION);
-
-    [
-        general_args,
-        process_args,
-        temperature_args,
-        cpu_args,
-        mem_args,
-        network_args,
-        #[cfg(feature = "battery")]
-        battery_args,
-        #[cfg(feature = "gpu")]
-        gpu_args,
-        style_args,
-        other_args,
-    ]
-    .into_iter()
-    .fold(cmd, |c, f| f(c))
+/// Returns an [`Command`] based off of [`BottomArgs`].
+#[cfg(test)]
+pub(crate) fn build_cmd() -> Command {
+    BottomArgs::command()
 }
 
 #[cfg(test)]
@@ -587,13 +585,13 @@ mod test {
 
     #[test]
     fn verify_cli() {
-        build_app().debug_assert();
+        build_cmd().debug_assert();
     }
 
     #[test]
     fn no_default_help_heading() {
-        let mut app = build_app();
-        let help_str = app.render_help();
+        let mut cmd = build_cmd();
+        let help_str = cmd.render_help();
 
         assert!(
             !help_str.to_string().contains("\nOptions:\n"),